Changed \rlap{?} to new macro \?
Previously \rlap{?} was used in tables to indicate questioned forms. This is hard to read when lots of forms are questioned. Moving to a superscript solves this, but then \rlap{textsuperscript*{?}} takes up lots of space.

Added a new macro \? which does the \rlap{} and the superscript question mark (with bold to compensate for artificial resizing of the glyph). The intermediate macro \supques does not do the \rlap{} so it can be used outside of tables. Used \DeclareDocumentCommand instead of \ProvideDocumentCommand because the existing \? (which does ¿) is in the way.